rename BuildingDialog to RentalDialog
