+++ Source/WPF/MyMoney/Dialogs/RentalDialog.xaml
@@ -1,4 +1,4 @@
-﻿<dialog:BaseDialog x:Class="Walkabout.Dialogs.BuildingDialog"
+﻿<dialog:BaseDialog x:Class="Walkabout.Dialogs.RentalDialog"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         mc:Ignorable="d" xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
@@ -80,7 +80,7 @@
                 <!-- Units -->
                 <TabItem Header="Units" >
                     <DataGrid x:Name="Units" Grid.Row="10" Grid.Column="0" Grid.ColumnSpan="4" ItemsSource="{Binding}"  
-                      Style="{StaticResource DefaultDataGridStyle}" 
+                              Style="{StaticResource DefaultDataGridStyle}"
                       MinHeight="100" AutoGenerateColumns="False">
                         <DataGrid.Columns>
                             <DataGridTextColumn Header="Number" Binding="{Binding Name}" />
@@ -113,7 +113,7 @@
                                   DisplayMemberPath="Name"  
                                   SelectedValuePath="Id" 
                                   SelectedValue="{Binding CategoryForIncome}"
-                                  ItemsSource="{Binding RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type dialog:BuildingDialog}}, Path=Categories}" HorizontalAlignment="Stretch"/>
+                                  ItemsSource="{Binding RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type dialog:RentalDialog}}, Path=Categories}" HorizontalAlignment="Stretch"/>
 
                         <TextBlock Text="Taxes :" Grid.Row="1" TextAlignment="Right" VerticalAlignment="Center"/>
                         <ComboBox Grid.Row="1" Grid.Column="2"
@@ -121,7 +121,7 @@
                                   DisplayMemberPath="Name"  
                                   SelectedValuePath="Id" 
                                   SelectedValue="{Binding CategoryForTaxes}"
-                                  ItemsSource="{Binding RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type dialog:BuildingDialog}}, Path=Categories}" HorizontalAlignment="Stretch"/>
+                                  ItemsSource="{Binding RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type dialog:RentalDialog}}, Path=Categories}" HorizontalAlignment="Stretch"/>
 
                         <TextBlock Text="Interest :" Grid.Row="2" TextAlignment="Right" VerticalAlignment="Center"/>
                         <ComboBox Grid.Row="2" Grid.Column="2"
@@ -129,7 +129,7 @@
                                   DisplayMemberPath="Name"  
                                   SelectedValuePath="Id" 
                                   SelectedValue="{Binding CategoryForInterest}"
-                                  ItemsSource="{Binding RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type dialog:BuildingDialog}}, Path=Categories}" HorizontalAlignment="Stretch"/>
+                                  ItemsSource="{Binding RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type dialog:RentalDialog}}, Path=Categories}" HorizontalAlignment="Stretch"/>
 
                         <TextBlock Text="Repairs :"  Grid.Row="3" TextAlignment="Right" VerticalAlignment="Center"/>
                         <ComboBox Grid.Row="3" Grid.Column="2"
@@ -137,7 +137,7 @@
                                   DisplayMemberPath="Name"  
                                   SelectedValuePath="Id" 
                                   SelectedValue="{Binding CategoryForRepairs}"
-                                  ItemsSource="{Binding RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type dialog:BuildingDialog}}, Path=Categories}" HorizontalAlignment="Stretch"/>
+                                  ItemsSource="{Binding RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type dialog:RentalDialog}}, Path=Categories}" HorizontalAlignment="Stretch"/>
 
                         <TextBlock Text="Maintenance :" Grid.Row="4" TextAlignment="Right" VerticalAlignment="Center"/>
                         <ComboBox Grid.Row="4" Grid.Column="2"
@@ -145,7 +145,7 @@
                                   DisplayMemberPath="Name"  
                                   SelectedValuePath="Id" 
                                   SelectedValue="{Binding CategoryForMaintenance}"
-                                  ItemsSource="{Binding RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type dialog:BuildingDialog}}, Path=Categories}" HorizontalAlignment="Stretch"/>
+                                  ItemsSource="{Binding RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type dialog:RentalDialog}}, Path=Categories}" HorizontalAlignment="Stretch"/>
 
                         <TextBlock Text="Management :" Grid.Row="5" TextAlignment="Right" VerticalAlignment="Center"/>
                         <ComboBox Grid.Row="5" Grid.Column="2"
@@ -153,7 +153,7 @@
                                   DisplayMemberPath="Name"  
                                   SelectedValuePath="Id" 
                                   SelectedValue="{Binding CategoryForManagement}"
-                                  ItemsSource="{Binding RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type dialog:BuildingDialog}}, Path=Categories}" HorizontalAlignment="Stretch"/>
+                                  ItemsSource="{Binding RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type dialog:RentalDialog}}, Path=Categories}" HorizontalAlignment="Stretch"/>
 
                     </Grid>
                 </TabItem>
